feat: create EmptyState component and refactor usage in CommunityPostList and other components
