+++ web/app/(support)/support/docs/_components/help-reader.tsx
@@ -13,7 +13,8 @@ import { SupportArticleContent } from "@/app/(support)/support/_components/suppo
 import { PublicArticleToc } from "@/app/(support)/support/_components/support-article-toc"
 import { SupportHeader } from "@/app/(support)/support/_components/support-header"
 import { SupportDocLink, type DocPageNavigationHandler, useSupportDocRoute } from "@/app/(support)/support/_components/support-help-navigation"
-import { SupportEmptyState as EmptyState, SupportSearchInput } from "@/app/(support)/support/_components/support-ui"
+import { SupportSearchInput } from "@/app/(support)/support/_components/support-ui"
+import { EmptyState } from "@/components/empty-state"
 import { useSupportDocReader as useSupportDocReaderData } from "@/app/(support)/support/_components/use-support-help-reader"
 import { useI18n } from "@/i18n/provider"
 import { fetchDocPages, submitDocPageFeedback, type DocPage } from "@/lib/api/support"
